backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-docx==0.8.11
python-multipart==0.0.6 